Double variables replaced by Floats in order to reduce Database memory cost, refactoring - variable, database table names changed.

+++ src/main/java/com/alexsoft/smarthouse/utils/TempUtils.java
@@ -2,7 +2,7 @@ package com.alexsoft.smarthouse.utils;
 
 public class TempUtils {
 
-    public Integer calculateRelativeHumidity(final Double temp, final Double devpoint) {
+    public Integer calculateRelativeHumidity(final Float temp, final Float devpoint) {
         if (temp == null || devpoint == null || devpoint > temp) {
             return null;
         }
@@ -13,12 +13,12 @@ public class TempUtils {
         return (int) rh;
     }
 
-    public Double calculateAbsoluteHumidity(Double temp, Integer rh) {
+    public Float calculateAbsoluteHumidity(Float temp, Integer rh) {
         if (temp == null || rh == null) {
             return null;
         }
         double ah = 6.112 * Math.pow(2.71828, 17.67 * temp / (243.5 + temp)) * rh * 2.1674 / (273.15 + temp);
-        return Math.round(ah * 10) / 10.0;
+        return Math.round(ah * 10) / 10.0F;
 
     }
 }
